refactor(packages/sui-pde): move server/client strategies to its own folder
